add skelMerge function to merge the skeleton into a chain
This patch does a packet merge on the generated skeleton packets with the
original chain. The skelMerge function does the right thing with respect to
putting all the packets in the right order. However, the pages are not
written in the correct order (all the skeleton pages are written first),
probably because all the skeleton packets are created with page index 0.
Consequently, the addskel subcommand currently produces just a packet dump of
the file that should be created. That packet dump is correct; the code for
then producing the final file is commented out.

darcs-hash:20061124120933-4b390-3c91918f3207931b6fbc73e5753b66bfee7dc9f5.gz