change the type of Timestamp from Rational to (Integer, Integer), in order
to keep the original denominator around without simplifying: don't lose it
as it is derived from the framerate / samplerate. This patch ensures that
frame offsets are printed as the remainder relative to the framerate. Prior
to this, dump timestamps were screwing up when the frame no and framerate
had a common divisor.

darcs-hash:20080225090911-4b390-cd9a84537d451332dd6143a7dd45509aa44114f2.gz